Initial timeout fd stuff. There is no code that uses it yet, but its there and running for when I (am about to) do.
The idea is that you register a callback to be called after a given time
period has elapsed. This is good for things like checking ping returns,
where you set the timeout period on the fd to be 2*ping, and when you get
a PONG you reset the timeout period. If the socket ever times out, you
get a ping timeout.

Its also going to be used for the up-coming non-blocking connect() code,
where a socket timeout kicks in to prevent a connect() taking too long
to run.

Once check_pings() has been split up, this and the poll/select/etc loop should
be the only two places that walking the entire fd_table[] array is allowed.


git-svn-id: http://svn.ratbox.org/svnroot/trunk@1298 b93f080a-4cfa-0310-a19a-80a68f531ef9