Made some changes to strncpy_irc().
Basically on systems using gcc with -O2 set, gcc will provide its own
builtin version of memcpy() which is far faster than strncpy_irc() so
we set a macro to do that.

For systems not using gcc, but have inlined functions, we provide an inlined
version of strncpy_irc().

Alas, for systems that neither use gcc or support inlined functions we fall
back to the plain old version we've been using all the long.


git-svn-id: http://svn.ratbox.org/svnroot/trunk@7937 b93f080a-4cfa-0310-a19a-80a68f531ef9