tiny format change in parse.c
The guts of it is here. When using sendto_match_butone() which is =only=
used to do the "oper magic" messages. i.e. /msg $server.mask and /msg #hostmask
The prefix sent to remote servers is only the nick, not the nick!user@host.
The prefix sent to local users is nick!user@host


git-svn-id: http://svn.ratbox.org/svnroot/trunk@9164 b93f080a-4cfa-0310-a19a-80a68f531ef9