Bring back my uintptr_t casts to the code which converts dlsym(3)'s return value to something we can make into a function pointer within the bounds of the C standard.  This is a hack as such, but it is far more "correct" than trying to convert two pointers to eachother, one being a data pointer (void*) and one being a function pointer.  Document this as such, as it is important to compilers like TenDRA, and also mention that I _do_ have something which is correct, but that for POLA sake, and since this will work pretty much everywhere, I am choosing not to put it in at this time.
Also, convert casts of socklen_t pointers from 'int*' to 'socklen_t*'.
Originally these were 'unsigned int*', which is correct on BSDish systems,
but may not be if for some reason socklen_t is signed (it isn't with BSD
networking), and really 'size_t foo; use(&foo)' doesn't need to be
'use((socklen_t *)&foo)', but it doesn't hurt anyway.  If this is causing
problems on some system, if it does not have socklen_t, we should make
socklen_t a signed integer on that system using our typedef'd version, as
it is infinitely more common for these parameters to be 'socklen_t*' which
is not the same as 'int*'.

The current source tree will now build happily on TenDRA systems again.


git-svn-id: http://svn.ratbox.org/svnroot/trunk@10333 b93f080a-4cfa-0310-a19a-80a68f531ef9