Sent in by Lee Approved by db
The patch does the following:
- fixes persistent channels to actually persist, we fixed them too far and
broke them
- removes the perm crap from persistent channels.  we should always
persist, no matter whether a user parts, quits or falls down a rabbit hole
into wonderland
- adds ircd-hybrid-7b14 stuff to RELNOTES
- adds docs about the $$ and $# stuff to whats-new, as well as updates
- adds the CHARSET=rfc1459 to 005

Theres still issues with persistent channels, but im too tired atm to do
anything.  Basically - the event system/cleanup_channels() doesnt cope with
the fact a channel should persist for at least persist_time.  It uses the
TS of the channel - so if a user parts a 1 year old channel, itll
persist.  If the cleanup_channels() event is called 10s later, itll be
destroyed in that event - we need to add a chptr->ts_persist to mark when
we started persisting, and use that instead of TS in cleanup_channels().


git-svn-id: http://svn.ratbox.org/svnroot/trunk@10542 b93f080a-4cfa-0310-a19a-80a68f531ef9