support commio-based events.
some i/o mechanisms (at least ports and kqueue) allow the user to queue
timers as well as i/o events.  by using this instead of our own event loop,
we can ask the OS to schedule events for us and avoid the cost of polling
the event table.  only implemented for ports at the moment.


git-svn-id: http://svn.ratbox.org/svnroot/trunk@18991 b93f080a-4cfa-0310-a19a-80a68f531ef9